axfi/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data handling
pandas>=2.2.0
numpy>=1.26.0
duckdb>=0.9.0
yfinance>=0.2.33
requests>=2.31.0

# Live data providers
alpaca-trade-api>=2.0.0
polygon-api-client>=1.0.0

# Analytics and ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
scipy>=1.11.0

# Technical analysis
ta>=0.11.0
pandas-ta>=0.3.14b

# Visualization
plotly>=5.18.0

# Scheduling & automation
APScheduler>=3.10.0

# ML utilities
joblib>=1.3.0
shap>=0.43.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
